fix raw parameters

+++ feature/detail/src/main/java/com/buntupana/tmdb/feature/detail/data/raw/TvShowDetailsRaw.kt
@@ -52,7 +52,7 @@ data class TvShowDetailsRaw(
     @Json(name = "vote_average")
     val voteAverage: Double?,
     @Json(name = "vote_count")
-    val voteCount: Long?,
+    val voteCount: Int?,
     val videos: MediaVideosRaw?,
     @Json(name = "content_ratings")
     val contentRatings: ContentRatingsRaw?,
